backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
SQLAlchemy==2.0.38
psycopg2-binary==2.9.10
python-dotenv==1.0.1
httpx==0.28.1
python-jose[cryptography]==3.3.0
bcrypt==5.0.0
pydantic[email]==2.10.6
